change objective force
I have changed the objective force to the one Emek suggested. I found it is much more simple, as long as I use the objective direction.

One bug I found is that in person to person force, if two people get too close, it is possible that one of these guys will be bounced out to very far away due to division over a very small distance.